requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI / LLM
anthropic>=0.40.0
voyageai>=0.3.0

# Database
supabase>=2.0.0

# Scraping
apify-client>=1.7.0

# Utilities
httpx>=0.27.0
beautifulsoup4>=4.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
